requirements.txt
streamlit
openai>=1.30.0
psycopg2-binary
python-dotenv